Fix up CursorLineNr
After the previous fixes, the CursorLineNr was a bit too dark.
The dark colour could be seen with:

set cursorline
set number

This change aligns the CursorLineNr with the old look-and-feel and
fine-tunes it to also be slightly duller than before. The change applies
both to the UI and console colors.

Signed-off-by: Jani Nurminen <slinky@iki.fi>