Add initial :terminal coloring
The terminal was not colored like the rest of the Vim, change them to be
close to colors used by Zenburn.

Right now the color order is as usually used in the terminal (black,
red, green, yellow, ...). It might make sense to tint the terminal
towards one or two colors but that shall be for future work.

Coloring the terminal Vim seems to need ":set termguicolors".

Terminal coloring feature works in at least Vim 8.2.0+ but seems
generally buggy; I've not seen the high 7 colors used.

Fixes #30

Signed-off-by: Jani Nurminen <slinky@iki.fi>